requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
fastapi
uvicorn
langchain
openai
requests
beautifulsoup4
playwright
python-dotenv
PyPDF2
python-docx